fix(charging): prevent smallint overflow in energy calculation and recalc migration on battery-side DC readings (#5617)
charger_actual_current and charger_voltage are smallint columns, and PostgreSQL multiplies two smallints into a smallint. Vehicles that write battery-side DC readings into these fields produce products above 32,767 (e.g. 155 A × 230 V), which crashed the recalc migration at boot and — latently since the columns were narrowed in 2020 — the energy calculation at charging completion. Cast the first factor to integer in the migration, determine_phases/1, and the charge-details determine_phases variable; in calculate_energy_used/1 reordering the factors makes the chain float from the first product on.

The released migration is edited in place on purpose: installations where it crashed never recorded it and will run the corrected version, installations where it succeeded computed identical values and skip it.

+++ priv/repo/migrations/20260808090000_recalc_charge_energy_used.exs
@@ -11,7 +11,7 @@ defmodule TeslaMate.Repo.Migrations.RecalcChargeEnergyUsed do
     execute("""
     WITH detection AS (
       SELECT c.charging_process_id AS cp_id,
-             avg(c.charger_power * 1000.0 / nullif(c.charger_actual_current * c.charger_voltage, 0)) AS p,
+             avg(c.charger_power * 1000.0 / nullif(c.charger_actual_current::int * c.charger_voltage, 0)) AS p,
              avg(c.charger_phases)::int AS r,
              count(*) AS n
       FROM charges c
@@ -28,7 +28,7 @@ defmodule TeslaMate.Repo.Migrations.RecalcChargeEnergyUsed do
     ), integrated AS (
       SELECT c.charging_process_id AS cp_id,
              CASE WHEN c.charger_phases IS NULL THEN c.charger_power::float
-                  ELSE coalesce(c.charger_actual_current * c.charger_voltage * ph.ph / 1000.0, c.charger_power)
+                  ELSE coalesce(c.charger_actual_current::int * c.charger_voltage * ph.ph / 1000.0, c.charger_power)
              END *
              EXTRACT(epoch FROM (c.date - lag(c.date) OVER (PARTITION BY c.charging_process_id ORDER BY c.date))) / 3600 AS energy
       FROM charges c
